+++ data-access/tests/test_zarr.py
@@ -12,19 +12,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import s3fs
-from nexustiles.nexustiles import NexusTileService
-
 import configparser
-import io, os
+import io
+import os
 
-from moto import mock_s3
-
-import mock, json
-
-import pytest
-
+import json
+import mock
 import numpy as np
+import pytest
+import s3fs
+from moto import mock_s3
+from nexustiles.nexustiles import NexusTileService
 
 class DummyTile:
     def __init__(self, tile_id):
@@ -111,22 +109,6 @@ def bounds():
     }
 
 @pytest.fixture()
-def config():
-    cfg = f"""
-    [s3]
-    bucket={bucket_name}
-    key={root_key}
-    region={region}
-    public=false
-    """
-
-    buf = io.StringIO(cfg)
-    config = configparser.ConfigParser()
-    config.read_file(buf)
-
-    yield config
-
-@pytest.fixture()
 def tile_service(s3):
     from nexustiles.dao.ZarrProxy import ZarrProxy
 
@@ -173,7 +155,6 @@ def test_bounds(bounds, tile_service):
 
 def test_matchup(bounds, tile_service):
     import webservice.algorithms_spark.Matchup as matchup
-    from webservice.algorithms.doms import insitu
 
     tile_service_factory = mock.MagicMock()
     tile_service_factory.return_value = tile_service
@@ -190,7 +171,6 @@ def test_matchup(bounds, tile_service):
     tiles = [tile_id]
 
     with mock.patch('webservice.algorithms_spark.Matchup.edge_endpoints.getEndpointByName') as mock_edge_endpoints:
-
         match_args = dict(
             tile_ids=tiles,
             primary_b=MockSparkParam('MUR25-JPL-L4-GLOB-v04.2'),
@@ -199,7 +179,7 @@ def test_matchup(bounds, tile_service):
             tt_b=MockSparkParam(43200),
             rt_b=MockSparkParam(1000),
             platforms_b=MockSparkParam('42'),
-            bounding_wkt_b=MockSparkParam('POLYGON((20 -100, 30 -100, 30 -79, 20 -79, 20 -100))'),
+            bounding_wkt_b=MockSparkParam('POLYGON((-100 20, -79 20, -79 30, -100 30, -100 20))'),
             depth_min_b=MockSparkParam(-20.0),
             depth_max_b=MockSparkParam(10.0),
             tile_service_factory=tile_service_factory
@@ -224,8 +204,6 @@ def test_matchup(bounds, tile_service):
 
             return filtered
 
-        result = filter_time(result)
-
         # [(k,v),...] -> {k: [v,...],...}
         def to_map(res):
             mapped = {}
@@ -241,6 +219,49 @@ def test_matchup(bounds, tile_service):
 
             return mapped
 
-        print(result)
+        result = to_map(filter_time(result))
 
-        assert True
+        assert len(result) == 4
+
+        keys = list(result.keys())
+
+        def validate_point(point, time, lon, lat, value, name, cf_name=None, secondary_point=False):
+            assert point.time == time
+            assert point.longitude == lon
+            assert point.latitude == lat
+            if not secondary_point:
+                assert point.data[0].variable_value == value
+                assert point.data[0].variable_name == name
+                assert point.data[0].cf_variable_name == cf_name
+            else:
+                assert point.data[0].variable_value == value[0]
+                assert point.data[0].variable_name == name[0]
+                assert point.data[1].variable_value == value[1]
+                assert point.data[1].variable_name == name[1]
+
+        validate_point(keys[0], '2018-08-17T09:00:00Z', -90.125, 27.625, 303.447998046875, 'analysed_sst', 'sea_surface_foundation_temperature')
+        validate_point(keys[1], '2018-08-21T09:00:00Z', -90.375, 28.125, 303.49200439453125, 'analysed_sst', 'sea_surface_foundation_temperature')
+        validate_point(keys[2], '2018-08-22T09:00:00Z', -90.125, 28.125, 303.3800048828125, 'analysed_sst', 'sea_surface_foundation_temperature')
+        validate_point(keys[3], '2018-08-27T09:00:00Z', -86.125, 27.625, 303.4729919433594, 'analysed_sst', 'sea_surface_foundation_temperature')
+
+        v0 = result[keys[0]]
+        v1 = result[keys[1]]
+        v2 = result[keys[2]]
+        v3 = result[keys[3]]
+
+        assert len(v0) == 6
+        assert len(v1) == 1
+        assert len(v2) == 1
+        assert len(v3) == 2
+
+        validate_point(v0[0], '2018-08-17T05:00:00Z', -90.13, 27.62, [30.4,1], ['sea_water_temperature', 'sea_water_temperature_quality'], secondary_point=True)
+        validate_point(v0[1], '2018-08-17T05:30:00Z', -90.13, 27.62, [30.4,1], ['sea_water_temperature', 'sea_water_temperature_quality'], secondary_point=True)
+        validate_point(v0[2], '2018-08-17T06:00:00Z', -90.13, 27.62, [30.4,1], ['sea_water_temperature', 'sea_water_temperature_quality'], secondary_point=True)
+        validate_point(v0[3], '2018-08-17T06:30:00Z', -90.13, 27.63, [30.4,1], ['sea_water_temperature', 'sea_water_temperature_quality'], secondary_point=True)
+        validate_point(v0[4], '2018-08-17T07:00:00Z', -90.13, 27.63, [30.4,1], ['sea_water_temperature', 'sea_water_temperature_quality'], secondary_point=True)
+        validate_point(v0[5], '2018-08-17T07:30:00Z', -90.13, 27.63, [30.3,1], ['sea_water_temperature', 'sea_water_temperature_quality'], secondary_point=True)
+        validate_point(v1[0], '2018-08-21T01:00:00Z', -90.38, 28.12, [30.0,1], ['sea_water_temperature', 'sea_water_temperature_quality'], secondary_point=True)
+        validate_point(v2[0], '2018-08-22T01:00:00Z', -90.13, 28.12, [30.3,1], ['sea_water_temperature', 'sea_water_temperature_quality'], secondary_point=True)
+        validate_point(v3[0], '2018-08-27T12:30:00Z', -86.12, 27.62, [30.0,1], ['sea_water_temperature', 'sea_water_temperature_quality'], secondary_point=True)
+        validate_point(v3[1], '2018-08-27T13:00:00Z', -86.13, 27.63, [30.0,1], ['sea_water_temperature', 'sea_water_temperature_quality'], secondary_point=True)
+